requirements.txt
wandb==0.21.3
monai==1.5.1
SimpleITK==2.5.2
numpy==2.1.2
pandas==2.3.3
scikit-image==0.25.2
scikit-learn==1.7.2
tqdm==4.67.1
matplotlib==3.10.7
hydra-core==1.3.2
timm==1.0.20
einops==0.8.1
lpips==0.1.4
nibabel==5.3.2
pydicom==3.0.1
rt_utils==1.2.7
torchdiffeq==0.2.5
